Protocols: add `spin_type` to `PwBaseWorkChain` input generator
This argument allows a caller to specify the spin polarization type that
is to be used for the system. Currently, only `NONE` and `COLLINEAR` are
supported. In the latter case, the initial magnetization will be
determined automatically if it is not explicitly defined through the
argument `initial_magnetization`. This should be mapping of kind names
onto the starting magnetization, which is a float between -1 and 1.